Tailwind styling generation changes
